+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.*;
 import roomescape.auth.domain.AuthInfo;
 import roomescape.global.annotation.LoginUser;
 import roomescape.reservation.controller.dto.*;
-import roomescape.reservation.service.MemberReservationService;
 import roomescape.reservation.service.ReservationService;
 import roomescape.reservation.service.WaitingReservationService;
 
@@ -20,14 +19,11 @@ import java.util.List;
 public class ReservationController {
 
     private final ReservationService reservationService;
-    private final MemberReservationService memberReservationService;
     private final WaitingReservationService waitingReservationService;
 
     public ReservationController(ReservationService reservationService,
-                                 MemberReservationService memberReservationService,
                                  WaitingReservationService waitingReservationService) {
         this.reservationService = reservationService;
-        this.memberReservationService = memberReservationService;
         this.waitingReservationService = waitingReservationService;
     }
 
@@ -38,14 +34,14 @@ public class ReservationController {
             @RequestParam(value = "dateFrom", required = false) LocalDate startDate,
             @RequestParam(value = "dateTo", required = false) LocalDate endDate
     ) {
-        return memberReservationService.findMemberReservations(
+        return reservationService.findMemberReservations(
                 new ReservationQueryRequest(themeId, memberId, startDate, endDate));
     }
 
     @PostMapping("/reservations")
     public ResponseEntity<ReservationResponse> create(@LoginUser AuthInfo authInfo,
                                                       @RequestBody @Valid ReservationRequest reservationRequest) {
-        ReservationResponse response = memberReservationService.createReservation(authInfo, reservationRequest);
+        ReservationResponse response = reservationService.createReservation(authInfo, reservationRequest);
         return ResponseEntity.created(URI.create("/reservations/" + response.reservationId())).body(response);
     }
